fix: move governance proposals route to dedicated governance module
Routes/analytics.ts was misused to serve governance proposal logic and
mounted under /governance path. Renamed to routes/governance.ts and
updated the import in api/routes.ts to match. Closes #298.
